services/api-gateway/pyproject.toml
[tool.poetry]
name = "api-gateway"
version = "0.1.0"
description = "API Gateway (FastAPI + Strawberry GraphQL)"
authors = ["faztyG <verselabproject@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
strawberry-graphql = "^0.280.0"
httpx = "^0.28.1"
grpcio = "^1.74.0"
orjson = "^3.11.2"
redis = "^6.4.0"
sqlalchemy = ">=2.0.43,<3.0.0"
alembic = ">=1.16.4,<2.0.0"
nats-py = "^2.11.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
